ref to func
