+++ frontend/src/components/Ingredient.tsx
@@ -6,7 +6,7 @@ import { Button, ButtonLink } from "@/components/Buttons"
 import { TextInput, selectTarget, CheckBox } from "@/components/Forms"
 import { hasSelection } from "@/utils/general"
 import { useDrop, useDrag } from "react-dnd"
-import { DragDrop } from "@/dragDrop"
+import { DragDrop, handleDndHover } from "@/dragDrop"
 
 const emptyField = ({
   quantity,
@@ -192,61 +192,7 @@ export function Ingredient(props: {
 
   const [, drop] = useDrop({
     accept: DragDrop.INGREDIENT,
-    hover: (_item, monitor) => {
-      if (!ref.current) {
-        return
-      }
-
-      // tslint:disable-next-line:no-unsafe-any
-      const dragIndex: number = monitor.getItem().index
-      const hoverIndex = props.index
-
-      // Don't replace items with themselves
-      if (dragIndex === hoverIndex) {
-        return
-      }
-
-      // Determine rectangle on screen
-      const el = ref.current
-      const hoverBoundingRect = el.getBoundingClientRect()
-
-      // Get vertical middle
-      const hoverMiddleY =
-        (hoverBoundingRect.bottom - hoverBoundingRect.top) / 2
-
-      // Determine mouse position
-      const clientOffset = monitor.getClientOffset()
-      if (clientOffset == null) {
-        return
-      }
-
-      // Get pixels to the top
-      const hoverClientY = clientOffset.y - hoverBoundingRect.top
-
-      // Only perform the move when the mouse has crossed half of the items height
-      // When dragging downwards, only move when the cursor is below 50%
-      // When dragging upwards, only move when the cursor is above 50%
-
-      // Dragging downwards
-      if (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) {
-        return
-      }
-
-      // Dragging upwards
-      if (dragIndex > hoverIndex && hoverClientY > hoverMiddleY) {
-        return
-      }
-
-      // Time to actually perform the action
-      props.move?.({ from: dragIndex, to: hoverIndex })
-
-      // Note: we're mutating the monitor item here!
-      // Generally it's better to avoid mutations,
-      // but it's good here for the sake of performance
-      // to avoid expensive index searches.
-      // tslint:disable-next-line:no-unsafe-any
-      monitor.getItem().index = hoverIndex
-    }
+    hover: handleDndHover({ ref, props })
   })
 
   const [{ isDragging }, drag, preview] = useDrag({
